chore: move to component
